Fix: Move fixtures into different namespace
