+++ test/Fixture/ClassUsesTrait/ClassUsingTrait.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2017 Andreas Möller.
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE file that was distributed with this source code.
+ *
+ * @link https://github.com/localheinz/test-util
+ */
+
+namespace Localheinz\Test\Util\Test\Fixture\ClassUsesTrait;
+
+final class ClassUsingTrait
+{
+    use ExampleTrait;
+}
