--- test/Unit/Fixture/ClassUsesTrait/ClassUsingTrait.php
@@ -1,19 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/**
- * Copyright (c) 2017 Andreas Möller.
- *
- * For the full copyright and license information, please view
- * the LICENSE file that was distributed with this source code.
- *
- * @link https://github.com/localheinz/test-util
- */
-
-namespace Localheinz\Test\Util\Test\Unit\Fixture\ClassUsesTrait;
-
-final class ClassUsingTrait
-{
-    use ExampleTrait;
-}
